get users for cli backup without getting info (#3207)
When the cli aggregates all users, it doesn't need to look up user info
for anyone, it only needs their id and PN. The rest of the info (such as
service support) will get looked up on a one-by-one basis in connector
during backup run.
